move all page file for separate section userPage , adminPage
