+++ frontend/src/pages/adminPages/system.tsx
@@ -1,0 +1,12 @@
+import React from "react";
+
+
+const System: React.FC = () => {
+ 
+
+    return ( 
+       <p></p>
+    );
+};
+
+export default System;
